LayoutTests/fast/mediastream/media-stream-track-source-failure-expected.txt

CONSOLE MESSAGE: A MediaStreamTrack ended due to a capture failure
CONSOLE MESSAGE: A MediaStreamTrack ended due to a capture failure


PASS Create stream 
PASS End video track only 
PASS End audio track only 
